% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djpr_ggiraph.R
\name{calc_girafe_height}
\alias{calc_girafe_height}
\title{Calculate the height of a ggiraph object}
\usage{
calc_girafe_height(height_percent, base_height = 400, dpi = 72)
}
\arguments{
\item{height_percent}{Height of object, as percentage of base_height}

\item{base_height}{Pixels; default 400}

\item{dpi}{DPI; default 72}
}
\description{
Calculate the height of a ggiraph object
}
