% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djpr_ggiraph.R
\name{calc_girafe_width}
\alias{calc_girafe_width}
\title{Return width in inches for ggiraph::girafe() objects}
\usage{
calc_girafe_width(width_percent, window_width, dpi = 72, max_width = 1140)
}
\arguments{
\item{width_percent}{Percentage of the standard plotting area that
should be taken up by the object}

\item{window_width}{Width of the row in which the object will be
placed. Given by \code{plt_change()$width}}

\item{dpi}{Dots per inch}

\item{max_width}{Maximum width of object in pixels}
}
\description{
Return width in inches for ggiraph::girafe() objects
}
