% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centred_row.R
\name{centred_row}
\alias{centred_row}
\title{Create a Shiny fluidRow with centred content
This is a convenience function intended for use in Shiny apps where you
want content centred, surrounded to left and white by whitespace, using
\code{shiny::fluidRow()} and \code{shiny::column()}.}
\usage{
centred_row(
  content,
  left_space = 2,
  right_space = left_space,
  max_width_px = 1140
)
}
\arguments{
\item{content}{The content you wish to place in the row, such as
\code{plotOutput("plot")}.}

\item{left_space}{Numeric. Default is \code{2}. Must be between 1 and 10
(inclusive). See \code{Details}.}

\item{right_space}{Numeric. Default is equal to \code{left_space}.
Must be between 1 and 10 (inclusive). See \code{Details}.}

\item{max_width_px}{Maximum width of content, in pixels.}
}
\value{
a \code{shiny.tag} object that creates a three-column fluidRow.
}
\description{
Create a Shiny fluidRow with centred content
This is a convenience function intended for use in Shiny apps where you
want content centred, surrounded to left and white by whitespace, using
\code{shiny::fluidRow()} and \code{shiny::column()}.
}
\details{
This function creates a \code{shiny::fluidRow()} then divides that
row into three columns. The left and right columns are whitespace. The
content goes in the centre column. Any space not allocated to the left or
right space is left for the content.

Units of measurement for \code{left_space} and \code{right_space} are based on the
Bootstrap 12-wide grid system. A value of 2 therefore means that a column
will take up 1/6th of the horizontal space on the page.

The total width of each column must be at least 1. The total width of
the three columns must be equal to 12. Each column therefore cannot be
greater than 10 width.
}
\examples{

centred_row(shiny::plotOutput("plot1"))
}
