% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djpr_dash_theme.R
\name{djpr_dash_theme}
\alias{djpr_dash_theme}
\title{DJPR dashboard theme}
\usage{
djpr_dash_theme(...)
}
\arguments{
\item{...}{unused}
}
\description{
DJPR dashboard theme
}
\details{
add anywhere in a shinydashboard body to add DJPR theme and
ggiraph_js()
javascript
}
