% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djpr_dashboard.R
\name{djpr_dashboard}
\alias{djpr_dashboard}
\title{Title}
\usage{
djpr_dashboard(header, sidebar, body, title = NULL)
}
\arguments{
\item{header}{temp}

\item{sidebar}{temp}

\item{body}{temp}

\item{title}{temp}
}
\description{
Title
}
