% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djpr_fonts.R
\name{djpr_fonts}
\alias{djpr_fonts}
\title{DJPR font css}
\usage{
djpr_fonts(...)
}
\arguments{
\item{...}{unused}
}
\value{
HTML link to djpr font css
}
\description{
DJPR font css
}
\details{
Replaces default fonts with DJPR fonts
}
