% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djpr_h2_box.R
\name{djpr_h2_box}
\alias{djpr_h2_box}
\title{Dashboard title box}
\usage{
djpr_h2_box(
  heading,
  ...,
  width = 12,
  colour = "#201547",
  text_colour = "#FFFFFF"
)
}
\arguments{
\item{heading}{text for h2 heading}

\item{...}{additional argumnents passed to \code{shinydashboard::valueBox}}

\item{width}{width of box on 1 to 12 scale (bootstrap grid)}

\item{colour}{Colour for the box as css spesification}

\item{text_colour}{colour for text as css spesification}
}
\value{
HTML element
}
\description{
Dashboard title box
}
