% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djpr_plot_async.R
\name{djpr_plot_async}
\alias{djpr_plot_async}
\title{Asynchronous DJPR plot server}
\usage{
djpr_plot_async(id, plot_fun, ...)
}
\arguments{
\item{id}{output namespace id to be used in conjunction with \code{djpr_plot_ui}
or \code{djpr_plot_box}}

\item{plot_fun}{a function which generates a ggplot}

\item{...}{arguments passed to plot_fun}
}
\value{

}
\description{
Asynchronous DJPR plot server
}
