% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djpr_plot_box.R
\name{djpr_plot_box}
\alias{djpr_plot_box}
\title{Output plots as shinydashboard boxes}
\usage{
djpr_plot_box(id, interactive = TRUE, plot_height = "400px", width = 6, ...)
}
\arguments{
\item{id}{a Shiny \code{outputId} specific to the individual plot.}

\item{interactive}{Boolean; does this plot use ggiraph}

\item{plot_height}{height in css spesification}

\item{...}{additional html placed after the plot}
}
\value{
HTML UI component
}
\description{
Output plots as shinydashboard boxes
}
