% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djpr_plot_text.R
\name{djpr_plot_title}
\alias{djpr_plot_title}
\alias{djpr_plot_subtitle}
\alias{djpr_plot_caption}
\title{Functions to create title, subtitle, and caption of plots
Creates container
Currently implemented via inline CSS.}
\usage{
djpr_plot_title(..., rel_font_size = 150)

djpr_plot_subtitle(..., rel_font_size = 100)

djpr_plot_caption(..., rel_font_size = 73)
}
\arguments{
\item{...}{arguments passed to container (eg. \code{div})}

\item{rel_font_size}{Numeric. Relative font size. In percent, so \code{100} would
set the font size equal to the base font size.}
}
\value{
A \code{shiny.tag} object.
}
\description{
Functions to create title, subtitle, and caption of plots
Creates container
Currently implemented via inline CSS.
}
\examples{

djpr_plot_title("A title goes here")

djpr_plot_subtitle("A subtitle goes here", rel_font_size = 100)

# Use as container in shiny::textOutput()
\dontrun{
textOutput("title", container = djpr_plot_title)
}
}
