% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djpr_shiny_plot.R
\name{djpr_plot_ui}
\alias{djpr_plot_ui}
\title{Function creates a plot environment containing (optional)
title and subtitle, chart, notes, and download buttons
Takes as input a function to create a ggplot2 or ggirafe object}
\usage{
djpr_plot_ui(id)
}
\arguments{
\item{id}{a Shiny \code{outputId} specific to the individual plot.}
}
\value{
A \code{shiny.tag} object creating a plot environment, with
labels (title, subtitle, caption) as HTML text, a download button,
and optional input controls.
}
\description{
Function creates a plot environment containing (optional)
title and subtitle, chart, notes, and download buttons
Takes as input a function to create a ggplot2 or ggirafe object
}
\details{
To be used in conjunction with \code{djpr_plot_server()} Shiny module,
which provides the server-side outputs that \code{djpr_plot_ui()} expects. See
example.
}
\examples{
\dontrun{

library(shiny)
library(ggplot2)

ui <- fluidPage(
  djpr_plot_ui("plot")
)

plot_function <- function(data = economics,
                          title = "This is a title",
                          subtitle = "This is a subtitle",
                          caption = "This data comes from the ggplot2 package") {
  data \%>\%
    ggplot(aes(x = date, y = unemploy)) +
    geom_line() +
    labs(
      title = title,
      subtitle = subtitle,
      caption = caption
    ) +
    theme_minimal(base_size = 16)
}

server <- function(input, output, session) {
  djpr_plot_server("plot",
    plot_function,
    date_slider = TRUE
  )
}

shinyApp(ui, server)
}

}
