% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djpr_shiny_theme.R
\name{djpr_shiny_theme}
\alias{djpr_shiny_theme}
\title{Modify appearance of Shiny apps consistent with DJPR style}
\usage{
djpr_shiny_theme(...)
}
\arguments{
\item{...}{arguments passed to \code{bslib::bs_theme()}}
}
\description{
This function uses the \code{bslib} package to create a Shiny theme.
}
\examples{
ui <- shiny::navbarPage(title = "Example app!",
                theme = djpr_shiny_theme()
)

}
