% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djpr_tab_panel.R
\name{djpr_tab_panel}
\alias{djpr_tab_panel}
\title{Creates a \code{shiny::tabPanel()} that includes a table of contents,
using \code{toc_row()}}
\usage{
djpr_tab_panel(title, ..., toc_space = 2, right_space = toc_space)
}
\arguments{
\item{title}{The title of the panel}

\item{...}{The content of the panel, passed to \code{toc_row()}}

\item{toc_space}{Numeric; width (out of a total of 12) to be given to the
left area that incldues the table of contents}

\item{right_space}{Numeric; width for the right whitespace; default is
the same as \code{toc_space}}
}
\value{
A \code{shiny.tag} object; a \code{tabPanel}.
}
\description{
Creates a \code{shiny::tabPanel()} that includes a table of contents,
using \code{toc_row()}
}
\details{
Intended for use inside \code{djpr_page()}.
}
\examples{
\dontrun{
library(shiny)

example_panel <- djpr_tab_panel(title = "Example",
                               h1("A heading"),
                               "Lorem ipsum",
                               h2("A subheading"),
                               "Lorem ipsum",
                               h1("Another heading"))

ui <- fluidPage(
 djpr_page(
   title = "Example",
   example_panel
 )
)

server <- function(input, output, session) {

}

shinyApp(ui, server)
}

}
