% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djpr_with_spinner.R
\name{djpr_with_spinner}
\alias{djpr_with_spinner}
\title{Wrapper around shinycssloaders::withSpinner() with DJPR defaults}
\usage{
djpr_with_spinner(
  ui_element,
  type = 8,
  colour = "#2A6FA2",
  size = 0.8,
  hide.ui = TRUE,
  ...
)
}
\arguments{
\item{ui_element}{An element such as \code{plotOutput()}}

\item{type}{See \url{https://daattali.com/shiny/shinycssloaders-demo}}

\item{colour}{Hex code for spinner}

\item{size}{Numeric; Size relative to default (1 = default)}

\item{hide.ui}{Logical. When \code{TRUE}, spinner is shown over a blank space;
when \code{FALSE}, spinner is shown over previously output UI.}

\item{...}{arguments passed to \code{shinycssloaders::withSpinner()}}
}
\description{
This function displays a spinner in place of Shiny UI elements while
they are loading / refreshing
}
\details{
Use within UI script
}
\examples{
\dontrun{
plotOutput("plot") \%>\% djpr_with_spinner()
}
}
