% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djprshiny-package.R
\docType{package}
\name{djprshiny-package}
\alias{djprshiny}
\alias{djprshiny-package}
\title{djprshiny: Common Functions For DJPR Shiny Apps}
\description{
Provides some common functions used in Shiny apps made by the Victorian Department of Jobs, Precincts and Regions.
}
\author{
\strong{Maintainer}: Matt Cowgill \email{matt.cowgill@ecodev.vic.gov.au} (\href{https://orcid.org/0000-0003-0422-3300}{ORCID})

Other contributors:
\itemize{
  \item Department of Jobs, Precincts and Regions (Victoria) [copyright holder, funder]
}

}
\keyword{internal}
