% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dropdown.R
\name{download_ui}
\alias{download_ui}
\title{Create a button that drops down into two sub-options}
\usage{
download_ui(id, ...)
}
\arguments{
\item{id}{Shiny module id}

\item{...}{arguments passed to \code{shinyWidgets::dropdownButton()}}
}
\description{
Used to download chart or chart data. Intended as the UI side of a
a Shiny module.
}
