% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focus_box.R
\name{focus_box}
\alias{focus_box}
\title{Create a focus box}
\usage{
focus_box(...)
}
\arguments{
\item{...}{content for the box}
}
\description{
A focus box is a div() with a small amount of drop shadow and spacing
}
