% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djpr_heading_box.R
\name{headings}
\alias{headings}
\alias{djpr_h1_box}
\alias{djpr_h2_box}
\alias{djpr_h3_box}
\alias{djpr_h4_box}
\title{DJPR title boxes}
\usage{
djpr_h1_box(
  heading,
  ...,
  width = 12,
  colour = "#1F1547",           # djpr_blue
  text_colour = "#FFFFFF"
)

djpr_h2_box(
  heading,
  ...,
  width = 12,
  colour = "#62BB46",           # djpr_green
  text_colour = "#FFFFFF"
)

djpr_h3_box(
  heading,
  ...,
  width = 12,
  colour = "#C0E4B5",           # djpr_faint_green
  text_colour = "#FFFFFF"
)

djpr_h4_box(
  heading,
  ...,
  width = 12,
  colour = "#BCD3EF",
  text_colour = "#FFFFFF"       # djpr_faint_blue
)
}
\arguments{
\item{heading}{text for heading}

\item{...}{additional contents for the box}

\item{width}{width of box on 1 to 12 scale (bootstrap grid)}

\item{colour}{Colour for the box as css specification}

\item{text_colour}{colour for text as css specification}
}
\value{
HTML element
}
\description{
DJPR title boxes
}
