% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_labs.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{extract_labs}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{djprtheme}{\code{\link[djprtheme]{extract_labs}}}
}}

