% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{round_dates}
\alias{round_dates}
\alias{date_floor}
\alias{date_ceiling}
\title{Round dates within the month}
\usage{
date_floor(dates)

date_ceiling(dates)
}
\arguments{
\item{dates}{A vector of date(s)}
}
\value{
A vector of date(s) the same length as the input \code{dates}.
}
\description{
Given a vector of date(s), return the first or last day of the month in
which the date occurs
}
\details{
For each element in \code{dates}, \code{date_floor()} returns a date
corresponding to the first day of the month in which the \code{dates} element
falls. \code{date_ceiling()} returns the final day of the month.
}
\examples{
date_floor(Sys.Date())
date_floor(c(Sys.Date(), as.Date("2017-10-18")))
date_ceiling(c(Sys.Date(), as.Date("2017-10-18")))
}
