% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_active.R
\name{text_active}
\alias{text_active}
\title{Function to create 'active text'}
\usage{
text_active(string, numbers, colour = "#71c5e8", alpha = 25, extra_style = "")
}
\arguments{
\item{string}{A length-one character vector, with XX inserted at points where
you want the elements of \code{numbers} to be inserted}

\item{numbers}{A vector (numeric or character) of numbers and/or words you
want to insert at the points marked XX in \code{string}.}

\item{colour}{Background colour for the inserted text (\code{numbers})}

\item{alpha}{Alpha (transparency) for the inserted text (\code{numbers}), where
100 is no transparency, 0 is full transparency.}

\item{extra_style}{Additional CSS to style the inserted text, such as
\code{"font-weight: bold;"}.}
}
\value{
Length-one character vector, with inline CSS.
}
\description{
Active text takes a string, inserts numbers and/or words at pre-defined
points in the string, and styles the inserted numbers/words with CSS. Active
text is intended for use in Shiny apps.
}
\examples{
string <- "The unemployment  rate went up by XX percentage points to XX per cent."
numbers <- c(0.2, 6.1)
text_active(string, numbers)
text_active(string, numbers, colour = "#007b4b", alpha = 0.5)
text_active(string, numbers,
  colour = "#007b4b", alpha = 0.5,
  extra_style = "font-weight: bold;"
)
}
