% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toc_header.R
\name{toc_header}
\alias{toc_header}
\title{Function returns tagList() of items to add
to a page to implement the Bootstrap Table of Contents
plugin.}
\source{
https://afeld.github.io/bootstrap-toc/#options
}
\usage{
toc_header()
}
\value{
A Shiny tagList intended to be added to an HTML page
}
\description{
Function returns tagList() of items to add
to a page to implement the Bootstrap Table of Contents
plugin.
}
