% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toc_row.R
\name{toc_row}
\alias{toc_row}
\title{Create a Shiny \code{fluidRow()} with three columns: table of contents,
content, and right space}
\usage{
toc_row(
  ...,
  page_id,
  page_title = "",
  toc_space = 2,
  right_space = toc_space,
  max_width_px = 1140
)
}
\arguments{
\item{...}{Elements to include within the main column.}

\item{page_id}{An ID that uniquely identifies the page on which the TOC row
will be placed. Can be any string without spaces, such as "indicators" or
"regional-unemployment.}

\item{page_title}{Title of the page. Can include spaces. Used for the
title of the table of contents.}

\item{toc_space}{Numeric. Minimum is 1, maximum is 10. Proportion of the
width of the page that should be given to the table of contents, where
the total width of the page = 12. Default is \code{3}.}

\item{right_space}{Numeric. Minimum is 1, maximum is 10. Proportion of the
width of the page that should be given to the whitespace at the right hand
side, where the total width of the page = 12. By default, same as \code{toc_space}.}

\item{max_width_px}{Maximum width, in pixels, of the main content column.}
}
\description{
Use in place of \code{shiny::fluidRow()}. Each page should have a maximum of one
\code{toc_row()} - to create rows within this overall row, use \code{shiny::fluidRow()}.
}
\details{
Note that widths of columns are currently hard-coded at page level;
changes to row widths in \code{toc_row} will not affect objects created with
\code{djpr_plot_server()}.
}
