% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_summarise.R
\name{ts_summarise}
\alias{ts_summarise}
\title{Summarise a data frame containing ABS time series}
\usage{
ts_summarise(df)
}
\arguments{
\item{df}{Data frame containing ABS time series}
}
\value{
Data frame with the following columns
\itemize{
\item {series_id}{Unique ABS time series ID}
\item {series}{Name of the ABS time series}
\item {latest_date}{Date of latest value}
\item {latest_value}{Latest value}
\item {d_year_abs}{Change between latest observation and prior year in absolute terms}
\item {d_year_perc}{Change between latest observation and prior year in percentage terms}
\item {d_period_abs}{Change between previous and latest period in absolute terms}
\item {d_period_perc}{Change between previous and latest period in percentage terms}
\item {latest_month}{Month of latest value}
\item {latest_year}{Year of latest value}
}
}
\description{
Summarise a data frame containing ABS time series
}
