% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djpr_dash_theme.R
\name{djpr_dashboard}
\alias{djpr_dashboard}
\title{DJPR themed dashboard}
\usage{
djpr_dashboard(header, sidebar, body, title = NULL)
}
\arguments{
\item{header}{shinydashboard dashboardHead}

\item{sidebar}{shinydashboard dashboardSidebar}

\item{body}{shinydashboard dashboardBody}

\item{title}{temp}
}
\description{
DJPR themed dashboard
}
