% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djpr_ggiraph.R
\name{djpr_girafe}
\alias{djpr_girafe}
\alias{ggiraph_js}
\title{Render ggiraph::girafe() object in a Shiny app and resize appropriately
with the browser window}
\source{
https://stackoverflow.com/questions/65267602/can-a-ggiraph-interactive-plot-be-the-size-of-the-window-in-r-shiny
and https://stackoverflow.com/questions/45191999/ggiraph-plot-doesnt-resize-to-fit-the-page and
https://stackoverflow.com/questions/36995142/get-the-size-of-the-window-in-shiny
}
\usage{
djpr_girafe(ggobj, height = 5, width = 6, ...)

ggiraph_js(col_widths = c(2, 8, 2))
}
\arguments{
\item{ggobj}{A ggplot2 object}

\item{height}{height in inches}

\item{width}{width in inches}

\item{...}{Additional options;
passed to \code{options} argument of \code{ggiraph::girafe}.}

\item{col_widths}{numeric vector of length 3; elements correspond to
Bootstrap widths of 3 columns. Must sum to 12.}
}
\value{
A \code{ggiraph::girafe()} object.
}
\description{
As a Shiny window  (html container) is resized,
ggplot2 objects (rendered using plotOutput())
change their widths while retaining their height. By default,
ggiraph::girafeOutput() objects don't resize in the same way - they maintain
their aspect ratio. These functions render these objects in a way that
resizes with the browser window.
}
\details{
Note that the title, subtitle, and caption are removed from \code{ggobj} using
\code{djprtheme::remove_labs()}.

\code{djpr_girafe()} should be used within \code{renderGirafe({})}, in place of
\code{ggiraph::girafe()}, within the \code{server} component of a Shiny app.

\code{ggiraph_js()} should be called within the Shiny UI, as in
\code{fluidPage(ggiraph_js())}.
}
\examples{
\dontrun{
library(shiny)
library(ggiraph)
library(ggplot2)

ui <- djpr_page(
  title = "Dashboard title",
  djpr_tab_panel(
    title = "Panel title",
    "Some text",
    "This is a ggplot2 object",
    plotOutput("plot1",
      width = "100\%"
    ),
    "This is a standard ggiraph object",
    girafeOutput("plot2"),
    "This is a ggiraph object that resizes with the window",
    girafeOutput("plot3")
  )
)

server <- function(input, output, session) {

  # First, we create a plot we can re-use:
  static_plot <- reactive({
    ggplot(mtcars, aes(x = wt, y = mpg)) +
      geom_point()
  })

  # This code generates a standard static ggplot
  output$plot1 <- renderPlot({
    static_plot()
  })

  # This code generates a standard ggiraph that doesn't re-size as we want

  output$plot2 <- renderGirafe({
    ggiraph::girafe(ggobj = static_plot())
  })

  # This code generates a ggiraph that resizes with the browser window
  output$plot3 <- renderGirafe({
    req(input$plt_change)
    p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
      ggiraph::geom_point_interactive(aes(tooltip = cyl))


    plt_change <- reactive(input$plt_change)
    plot_width <- calc_girafe_width(
      width_percent = 100,
      window_width = plt_change()$width,
      dpi = plt_change()$dpi
    )

    plot_height <- calc_girafe_height(
      height_percent = 100,
      window_height = plt_change()$height,
      dpi = plt_change()$dpi
    )

    djpr_girafe(p,
      width = plot_width,
      height = plot_height
    )
  })
}
shinyApp(ui, server)
}
library(ggplot2)
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  labs(title = "Title will be removed in girafe plot")

djpr_girafe(p, 5, 6)
}
\keyword{internal}
