% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djpr_page.R
\name{djpr_page}
\alias{djpr_page}
\title{Set up a \code{shiny::fluidPage()} with DJPR characteristics}
\usage{
djpr_page(..., title = NULL, theme = djpr_shiny_theme(), lang = "en")
}
\arguments{
\item{...}{content for the page}

\item{title}{Title for the page}

\item{theme}{By default, \code{djpr_shiny_theme()}}

\item{lang}{"en" for English}
}
\value{
Fluid page; object of class \code{shiny.tag.list}
}
\description{
This function should be used as a drop-in replacement for \code{shiny::fluidPage}.
It sets up a fluidPage with the DJPR Shiny theme, and some Javascript to enable
correct resizing of ggiraph objects.
}
