% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djpr_ts_linechart.R
\name{djpr_ts_linechart}
\alias{djpr_ts_linechart}
\title{Create a standard time series linechart}
\usage{
djpr_ts_linechart(
  data,
  col_var = .data$series,
  dot = TRUE,
  label = TRUE,
  label_num = round(.data$value, 1),
  y_labels = ggplot2::waiver(),
  hline = NULL
)
}
\arguments{
\item{data}{A dataframe containing time series data.
Your dataframe is presumed to include, at a minimum:
\itemize{
\item{"date"}{"A date column, of class \code{"Date"}}
\item{"value"}{"A value column, of class \code{numeric}, containing data to be shown on the y-axis}
\item{col_var}{"A variable to map to colour, which can be specified with the \code{col_var} argument."}
}}

\item{col_var}{Variable in \code{data} to map to the colour aesthetic;
default is \code{series}}

\item{dot}{Logical; \code{TRUE} by default. When \code{TRUE}, a filled dot will be shown on the
most recent data point.}

\item{label}{Logical; \code{TRUE} by default. When \code{TRUE}, a text label will be
added to the right of the most recent data point.}

\item{label_num}{Variable name (or expression) defining the label to be placed
on the chart. Default is \code{round(value, 1)}. Ignored if \code{label} is \code{FALSE}.}

\item{y_labels}{Supplied to the \code{labels} argument of
\code{ggplot2::scale_y_continuous()}}

\item{hline}{Numeric. If non-\code{NULL} (the default), a horizontal line will
be drawn at the data value given. eg. if \code{hline} = \code{0}, a \code{geom_hline()}
will be added at y = 0.}
}
\value{
A ggplot2 object
}
\description{
Create a standard time series linechart
}
\details{
If a column called 'tooltip' is present, it will be used as the
ggiraph tooltip; if not, one will be created.
}
\examples{
\dontrun{
library(tidyverse)
data <- djprdashdata::download_abs_ts("6202.0") \%>\%
  dplyr::filter(table_no == "6202005")

data <- data \%>\%
  dplyr::filter(
    grepl("Employment to population ratio", series),
    series_type == "Seasonally Adjusted"
  )

djpr_ts_linechart(data = data)
}

}
