% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dropdown.R
\name{download_dropdown}
\alias{download_dropdown}
\title{Create a button that drops down into two sub-options}
\usage{
download_dropdown(session_ns, ...)
}
\arguments{
\item{session_ns}{Value should be \code{session$ns}. This is supplied rather than
an \code{id} directly, as it enables nesting Shiny modules that generate UI.}

\item{...}{arguments passed to \code{shinyWidgets::dropdownButton()}}
}
\description{
Used to download chart or chart data. Intended as the UI side of a
a Shiny module.
}
