% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dropdown.R
\name{download_server}
\alias{download_server}
\title{Server side of download_ui Shiny module}
\usage{
download_server(id, plot)
}
\arguments{
\item{id}{Module id}

\item{plot}{A ggplot2 object}
}
\description{
Server side of download_ui Shiny module
}
