% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round2.R
\name{round2}
\alias{round2}
\title{Round numbers to specified number of digits
This function is a drop-in replacement for \code{base::round()}}
\source{
https://stackoverflow.com/questions/12688717/round-up-from-5
}
\usage{
round2(x, digits = 0)
}
\arguments{
\item{x}{a numeric vector of numbers to round}

\item{digits}{integer indicating the number of decimal places
to which \code{x} should be rounded.}
}
\description{
Unlike \code{base::round()}, it does NOT follow the
IEC 60559 standard by rounding numbers ending in .5 to the even
digit. Instead, 0.5 is rounded up.
}
\examples{
# This gives '2'
base::round(2.5)

# This gives 3
round2(2.5)
}
